work/src/servo_bot/graff_rag/requirements.txt
llama-index
llama-index-graph-stores-memgraph
llama-index-embeddings-ollama
llama-index-llms-ollama
llama-index-readers-file
python-dotenv
pymgclient
pypdf
